Aman-Srivastava/src/Programs/Crud_Operations/requirements.txt
Flask~=2.3.2
pydantic~=1.10.7
mysql-connector-python~=8.0.33